+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # QOwnNotes Changelog
 
+## 1.1.1
+- Pressing <kbd>Ctrl</kbd> + <kbd>⇧</kbd> + <kbd>F</kbd> to do a note search 
+  now leaves the distraction free mode
+
 ## 1.1
 - QOwnNotes now allows the usage of `# ` markdown headlines as note title
   (for [Issue #87](https://github.com/pbek/QOwnNotes/issues/87))